nuclear_checklist_backend/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
sqlalchemy==2.0.41
pymssql==2.3.7
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7

